refactor: move duplicate functions from vscodeignoreRequiredPatterns.ts to packageJsonUtils.ts
