+++ packages/eslint-local-rules/src/vscodeignoreContributesConflict.ts
@@ -7,25 +7,12 @@
 
 import type { Rule } from 'eslint';
 import { minimatch } from 'minimatch';
-import * as fs from 'node:fs';
 import * as pathModule from 'node:path';
+import { isRecord, readJsonRecord } from './packageJsonUtils';
 
 type ParsedLine = {
   readonly line: number;
   readonly pattern: string;
-};
-
-const isRecord = (value: unknown): value is Record<string, unknown> =>
-  typeof value === 'object' && value !== undefined && value !== null;
-
-const readJsonRecord = (filePath: string): Record<string, unknown> | undefined => {
-  if (!fs.existsSync(filePath)) return undefined;
-  try {
-    const parsed = JSON.parse(fs.readFileSync(filePath, 'utf-8'));
-    return isRecord(parsed) ? parsed : undefined;
-  } catch {
-    return undefined;
-  }
 };
 
 /**
